workers/ai_tagger/requirements.txt
pika==1.3.2
boto3==1.34.34
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
